srcdockerfiles/requirements.txt
PyJWT==2.1.0
Flask
pytest
coverage
aws-cdk-lib
tox
tox-gh-actions